✨feat(suspense): add more skelton view
